Term3/BDM-3035/demo_site/requirements.txt
Flask==3.0.0
flask_cors==4.0.0
flask_restful==0.3.10
flask_swagger_ui==4.11.1
numpy==1.26.4
pandas==2.0.3
tensorflow==2.16.1
opencv-python==4.9.0.80
matplotlib==3.7.1
pillow==9.5.0
